Renmace tools names and move packages
